+++ src/collection/constructors/createCollectionFromDirectory.test.ts
@@ -19,7 +19,7 @@ describe('createCollectionFromDirectory', () => {
             },
         );
         let pipelineFromCollection = await collection.getPipelineByUrl(
-            'https://promptbook.studio/examples/simple.book.md',
+            'https://promptbook.studio/examples/simple.book',
         );
 
         pipelineFromCollection = unpreparePipeline(pipelineFromCollection);
@@ -41,7 +41,7 @@ describe('createCollectionFromDirectory', () => {
             },
         );
         let pipelineFromCollection = await collection.getPipelineByUrl(
-            'https://promptbook.studio/examples/simple.book.md',
+            'https://promptbook.studio/examples/simple.book',
         );
 
         pipelineFromCollection = unpreparePipeline(pipelineFromCollection);
@@ -62,7 +62,7 @@ describe('createCollectionFromDirectory', () => {
             },
         );
         let pipelineFromCollection = await collection.getPipelineByUrl(
-            'https://promptbook.studio/examples/jokers.book.md',
+            'https://promptbook.studio/examples/jokers.book',
         );
 
         pipelineFromCollection = unpreparePipeline(pipelineFromCollection);
@@ -115,7 +115,7 @@ describe('createCollectionFromDirectory', () => {
                     isRecursive: false,
                 });
                 const pipelineFromCollection = await collection.getPipelineByUrl(
-                    'https://promptbook.studio/webgpt/write-website-content.book.md',
+                    'https://promptbook.studio/webgpt/write-website-content.book',
                 );
                 return pipelineFromCollection.title;
             })(),