+++ src/collection/constructors/createCollectionFromJson.test.ts
@@ -11,7 +11,7 @@ describe('createCollectionFromJson', () => {
             Show how to use a simple completion prompt
 
             -   PROMPTBOOK VERSION 1.0.0
-            -   PIPELINE URL https://promptbook.studio/examples/pipeline.book.md
+            -   PIPELINE URL https://promptbook.studio/examples/pipeline.book
             -   INPUT  PARAMETER {thing} Any thing to buy
             -   OUTPUT PARAMETER {response}
 
@@ -37,7 +37,7 @@ describe('createCollectionFromJson', () => {
         const pipeline = await compilePipeline(pipelineString);
         const collection = createCollectionFromJson(pipeline);
         const pipelineFromCollection = await collection.getPipelineByUrl(
-            'https://promptbook.studio/examples/pipeline.book.md',
+            'https://promptbook.studio/examples/pipeline.book',
         );
         expect(pipelineFromCollection).toEqual(await compilePipeline(pipelineString));
     });