+++ src/utils/editable/utils/stringifyPipelineJson.test.ts
@@ -5,7 +5,7 @@ import { importPipelineJsonAsString } from '../../../conversion/validation/_impo
 import { stringifyPipelineJson } from './stringifyPipelineJson';
 
 describe('how stringifyPipelineJson works', () => {
-    it('should work with markdown-knowledge.book.json', () =>
+    it('should work with markdown-knowledge.bookc', () =>
         expect(
             stringifyPipelineJson({
                 index: [
@@ -30,8 +30,8 @@ describe('how stringifyPipelineJson works', () => {
             ) + '\n',
         ));
 
-    it('should work with markdown-knowledge.book.json', () =>
-        expect(stringifyPipelineJson(importPipelineJson('26-markdown-knowledge.book.json'))).toBe(
-            importPipelineJsonAsString('26-markdown-knowledge.book.json'),
+    it('should work with markdown-knowledge.bookc', () =>
+        expect(stringifyPipelineJson(importPipelineJson('26-markdown-knowledge.bookc'))).toBe(
+            importPipelineJsonAsString('26-markdown-knowledge.bookc'),
         ));
 });